Fixed some minor issues
