+++ tdei-ui/src/routes/Maps/Map.css
@@ -954,7 +954,6 @@
   .mapboxgl-ctrl-geocoder--icon {
     display: inline-block;
     vertical-align: middle;
-    speak: none;
     fill: #757575;
     top: 15px;
   }
