+++ tdei-ui/src/routes/Maps/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Maps";
