+++ tdei-ui/src/routes/Stations/CreateUpdateStation.js
@@ -3,60 +3,60 @@ import { Button, Form } from "react-bootstrap";
 import Container from "../../components/Container/Container";
 import Layout from "../../components/Layout";
 import { getSelectedOrg } from "../../selectors";
-import MapBox from "./MapBox";
+import MapBox from "../Maps/MapBox";
 import { useDispatch, useSelector } from "react-redux";
-import useCreateService from "../../hooks/service/useCreateService";
+import useCreateStation from "../../hooks/station/useCreateStation";
 import { useAuth } from "../../hooks/useAuth";
 import { show } from "../../store/notification.slice";
 import { show as showModal } from "../../store/notificationModal.slice";
 import OrgList from "../../components/OrganisationList/OrgList";
 import { Formik, Field } from "formik";
 import * as yup from "yup";
-import useUpdateSevice from "../../hooks/service/useUpdateService";
+import useUpdateStation from "../../hooks/station/useUpdateStation";
+import { GET_STATIONS } from "../../utils";
 import { useQueryClient } from "react-query";
-import style from "./Maps.module.css";
+import style from "../Maps/Maps.module.css";
 import { useNavigate, useParams } from 'react-router-dom';
 import { GEOJSON } from '../../utils'
-import "./Map.css"
-import { GET_SERVICES } from "../../utils";
-import { getService } from "../../services";
+import "../Maps/Map.css"
+import { getStation } from "../../services";
 
-const CreateUpdateService = () => {
+const CreateUpdateStation = () => {
     const queryClient = useQueryClient();
     const dispatch = useDispatch();
     const navigate = useNavigate();
-    var data = {};
     const selectedOrg = useSelector(getSelectedOrg);  
     const idData = useParams();
     const [geoJson, setGeoJson] = useState(GEOJSON);
-    const [serviceData, setServiceData] = React.useState({
-        service_name: "",
+    const [stationData, setStationData] = React.useState({
         tdei_org_id: selectedOrg.tdei_org_id,
+        station_name: "",
         polygon: GEOJSON
-      });
+    });
     const { user } = useAuth();
    
     console.log("Params ",idData['id']);
     React.useEffect(()=>{
         if(idData['id'] !== undefined){
-            getService(idData['id']).then((services)=>{
-                console.log(services);
-                setServiceData(services.data[0]);
+            getStation(idData['id']).then((stations)=>{
+                console.log(stations);
+                setStationData(stations.data[0]);
+                setGeoJson(stations.data[0].polygon)
             })
         }
     },[idData]);
 
     const validationSchema = yup.object().shape({
         tdei_org_id: yup.string().required("Organization Name is required"),
-        service_name: yup.string().required("Service Name is required"),
-      });
+        station_name: yup.string().required("Station Name is required"),
+    });
 
     const onSuccess = (data) => {
-        console.log("sucessfully created", data);
-        queryClient.invalidateQueries({ queryKey: [GET_SERVICES] });
+        console.log("suucessfully created", data);
+        queryClient.invalidateQueries({ queryKey: [GET_STATIONS] });
         dispatch(
             showModal({
-                message: `Service ${serviceData?.tdei_service_id ? "updated" : "created"
+                message: `Station ${stationData?.tdei_station_id ? "updated" : "created"
                     } successfully.`,
             })
         );
@@ -66,43 +66,30 @@ const CreateUpdateService = () => {
         console.error("error message", err);
         dispatch(
             show({
-                message: `Error in ${serviceData?.tdei_service_id ? "updating" : "creating"
-                    } Service`,
+                message: `Error in ${stationData?.tdei_station_id ? "updating" : "creating"
+                    } station`,
                 type: "danger",
             })
         );
         navigate(-1);
     };
-    const { isLoading, mutate } = useCreateService({ onSuccess, onError });
-  const { isLoading: isUpdateLoading, mutate: updateService } = useUpdateSevice(
-    { onSuccess, onError }
-  );
+    const { isLoading, mutate } = useCreateStation({ onSuccess, onError });
+    const { isLoading: isUpdateLoading, mutate: updateStation } =
+        useUpdateStation({ onSuccess, onError });
 
-        // const handleCreateService = (values) => {
-        //     if (props.data?.tdei_service_id) {
-        //       updateService({
-        //         service_name: values.service_name,
-        //         tdei_service_id: props.data?.tdei_service_id,
-        //         tdei_org_id: values.tdei_org_id,
-        //       });
-        //     } else {
-        //       mutate(values);
-        //     }
-        //   };
-
-    const handleCreateService = (values) => {
-        if (serviceData?.tdei_service_id) {
-            updateService({
-                service_name: values.service_name? values.service_name : "",
-                tdei_service_id: serviceData?.tdei_service_id ? serviceData?.tdei_service_id : "",
+    const handleCreateStation = (values) => {
+        if (stationData?.tdei_station_id) {
+            updateStation({
+                station_name: values.station_name? values.station_name : "",
+                tdei_station_id: stationData?.tdei_station_id ? stationData?.tdei_station_id : "",
                 tdei_org_id: values.tdei_org_id? values.tdei_org_id : "",
                 polygon: geoJson.features.length === 0 ? GEOJSON : geoJson
             });
         } else {
             console.log("geoJson", geoJson);
             mutate({
-                service_name: values.service_name,
-                tdei_service_id: serviceData?.tdei_service_id,
+                station_name: values.station_name,
+                tdei_station_id: stationData?.tdei_station_id,
                 tdei_org_id: values.tdei_org_id,
                 polygon: geoJson
             });
@@ -124,9 +111,9 @@ const CreateUpdateService = () => {
     };
     const getHeader = () => {
         if (idData['id'] !== undefined) {
-            return "Edit Service";
+            return "Edit Station";
         } else {
-            return "Create New Service";
+            return "Create New Station";
         }
     };
     const handlePop = () => {
@@ -137,9 +124,9 @@ const CreateUpdateService = () => {
         setGeoJson(data);
     };
     const getJson = () => {
-        if (selectedOrg?.tdei_org_id && serviceData?.polygon !== undefined ){
-            return serviceData?.polygon;
-        }else if (!selectedOrg?.tdei_org_id && serviceData?.polygon === undefined ){
+        if (selectedOrg?.tdei_org_id && stationData?.polygon !== undefined ){
+            return stationData?.polygon;
+        }else if (!selectedOrg?.tdei_org_id && stationData?.polygon === undefined ){
             return GEOJSON;
         }
         return geoJson
@@ -148,8 +135,8 @@ const CreateUpdateService = () => {
     return (
         <Layout>
             <Formik
-                initialValues={serviceData}
-                onSubmit={handleCreateService}
+                initialValues={stationData}
+                onSubmit={handleCreateStation}
                 validationSchema={validationSchema}
                 enableReinitialize
             >
@@ -203,23 +190,23 @@ const CreateUpdateService = () => {
                                     )}
                                 </Form.Group>
                                 <Form.Group className="col-7 mb-3" controlId="name">
-                                    <Form.Label>Service Name</Form.Label>
+                                    <Form.Label>Station Name</Form.Label>
                                     <Form.Control
                                         type="text"
-                                        placeholder="Enter Service Name"
-                                        name="service_name"
+                                        placeholder="Enter Station Name"
+                                        name="station_name"
                                         onChange={handleChange}
                                         onBlur={handleBlur}
-                                        value={values.service_name}
-                                        isInvalid={touched.service_name && !!errors.service_name}
+                                        value={values.station_name}
+                                        isInvalid={touched.station_name && !!errors.station_name}
                                     />
                                     <Form.Control.Feedback type="invalid">
-                                        {errors.service_name}
+                                        {errors.station_name}
                                     </Form.Control.Feedback>
                                 </Form.Group>
                                 <div className={style.apiKey}>
-                                    <Form.Label>Service Location</Form.Label>
-                                    <MapBox isEdit = {serviceData?.tdei_service_id ? true : false} geojson={getJson()} onGeoJsonAdded={handleGeoJson} />
+                                    <Form.Label>Station Location</Form.Label>
+                                    <MapBox isEdit = {stationData?.tdei_station_id ? true : false} geojson={getJson()} onGeoJsonAdded={handleGeoJson} />
                                 </div>
                             </Container>
                         </Form>
@@ -231,4 +218,4 @@ const CreateUpdateService = () => {
 };
 
 
-export default CreateUpdateService;
+export default CreateUpdateStation;
